fix contextID type, move ipni in market
